requirements.txt
Django==2.2.2
djangorestframework==3.9.4
pillow==6.0.0
django-mathfilters==0.4.0
